+++ changelog.md
@@ -3,8 +3,11 @@
 ## Changelog
 
 ### Version 5.1 - 2024.07.10
-* 
+
 * New: added several more icons to the icon font
+* Edit: updated links format for the Dev4Press website
+* Edit: few tweaks to the Pixabay integration implementation
+* Edit: few tweaks to the Words generator class
 
 ### Version 5.0 - 2024.07.10